reform/split resolvers files
