+++ graphql/resolvers/auth.js
@@ -2,6 +2,10 @@
 
 const bcrypt = require('bcryptjs');
 const User = require('../../models/user');
+const Booking = require('../../models/booking');
+const Event = require('../../models/event');
+const { dateToString } = require('../../helpers/date');
+const { events, singleEvent, user } = require('./util');
 
 module.exports = {
   createUser: async (args) => {
