+++ graphql/resolvers/util.js
@@ -1,5 +1,31 @@
 'use strict';
 
+const User = require('../../models/user');
+const Booking = require('../../models/booking');
+const Event = require('../../models/event');
+
+const { dateToString } = require('../../helpers/date');
+
+const transformBooking = booking => {
+  return {
+    ...booking._doc,
+    _id: booking.id,
+    user: user.bind(this, booking._doc.user),
+    event: singleEvent.bind(this, booking._doc.event),
+    createdAt: dateToString(booking._doc.createdAt),
+    updatedAt: dateToString(booking._doc.updatedAt)
+  };
+};
+
+const transformEvent = event => {
+  return {
+    ...event._doc,
+    _id: event.id,
+    date: dateToString(event._doc.date),
+    creator: user.bind(this, event.creator)
+  };
+};
+
 const events = async eventIds => {
   try {
     const events = await Event.find({_id: {$in: eventIds} });
@@ -37,3 +63,5 @@ const user = async (userId) => {
 exports.events = events;
 exports.singleEvent = singleEvent;
 exports.user = user;
+exports.transformEvent = transformEvent;
+exports.transformBooking = transformBooking;
